feat: show map from accordion form list
